embed/requirements.txt
flask
waitress
sentence_transformers
faiss-cpu
python-dotenv
pymongo
